+++ packages/manager-interface/src/legacy/containers/ClaimRewardCompetition.js
